refactor: Player 에서 Name 제거
사용자 이름까지 추상화하는 것은 너무 과한 추상화라고 판단하여
제거한다.
